Split up prism module and get tests to pass
