+++ cidc_schemas/prism/core.py
@@ -1,37 +1,17 @@
-import re
-import sys
-import json
-import copy
-import uuid
-import datetime
-import logging
-from typing import (
-    Union,
-    BinaryIO,
-    Tuple,
-    List,
-    NamedTuple,
-    Any,
-    Optional,
-    Dict,
-    Callable,
-)
+"""Build metadata dictionaries from Excel files."""
 
-import openpyxl
-import jsonschema
-from jsonmerge import merge, Merger, exceptions as jsonmerge_exceptions, strategies
-from collections import namedtuple
-from jsonpointer import JsonPointer, JsonPointerException, resolve_pointer, EndOfList
+import json
+import logging
+from typing import Any, List, NamedTuple, Tuple, Union
 
 from cidc_schemas.json_validation import load_and_validate_schema
 from cidc_schemas.template import Template
-from cidc_schemas.template_writer import RowType
 from cidc_schemas.template_reader import XlTemplateReader
+from cidc_schemas.template_writer import RowType
+from jsonmerge import Merger, strategies, exceptions as jsonmerge_exceptions
+from jsonpointer import EndOfList, JsonPointer, JsonPointerException, resolve_pointer
 
-from cidc_schemas.constants import SCHEMA_DIR, TEMPLATE_DIR
-from .util import get_path, get_source, load_pipeline_config_template
-
-PROTOCOL_ID_FIELD_NAME = "protocol_identifier"
+from .constants import SUPPORTED_TEMPLATES
 
 logger = logging.getLogger(__file__)
 
@@ -227,38 +207,11 @@ def __jpointer_insert_next_thing(doc, jpoint, part, next_thing):
             doc.append(next_thing)
 
 
-SUPPORTED_ASSAYS = [
-    "wes_fastq",
-    "wes_bam",
-    "olink",
-    "cytof",
-    "ihc",
-    "elisa",
-    "rna_fastq",
-    "rna_bam",
-]
-
-SUPPORTED_SHIPPING_MANIFESTS = [
-    "pbmc",
-    "plasma",
-    "tissue_slide",
-    "normal_blood_dna",
-    "normal_tissue_dna",
-    "tumor_tissue_dna",
-    "h_and_e",
-]
-# weird non shipping manifest
-SUPPORTED_WEIRD_MANIFESTS = ["tumor_normal_pairing"]
-SUPPORTED_MANIFESTS = SUPPORTED_SHIPPING_MANIFESTS + SUPPORTED_WEIRD_MANIFESTS
-
-SUPPORTED_ANALYSES = ["cytof_analysis", "wes_analysis"]
-
-SUPPORTED_TEMPLATES = SUPPORTED_ASSAYS + SUPPORTED_MANIFESTS + SUPPORTED_ANALYSES
-
-LocalFileUploadEntry = namedtuple(
-    "LocalFileUploadEntry",
-    ["local_path", "gs_key", "upload_placeholder", "metadata_availability"],
-)
+class LocalFileUploadEntry(NamedTuple):
+    local_path: str
+    gs_key: str
+    upload_placeholder: str
+    metadata_availability: bool
 
 
 def _process_property(
@@ -770,507 +723,3 @@ def prismify(
         root_ct_obj = template_root_obj
 
     return root_ct_obj, collected_files, errors_so_far
-
-
-def _set_data_format(ct: dict, artifact: dict):
-    """
-    Discover the correct data format for the given artifact.
-    Args:
-        ct: a clinical trial object with artifact inserted
-        artifact: a reference to the artifact object inserted in `ct`.
-        NOTE: in-place updates to artifact must trigger in-place updates to `ct`.
-    """
-    # This is invalid for all artifact types, and will
-    # deliberately trigger a validation error below.
-    artifact["data_format"] = "[NOT SET]"
-
-    validator: jsonschema.Draft7Validator = load_and_validate_schema(
-        "clinical_trial.json", return_validator=True
-    )
-
-    for error in validator.iter_errors(ct):
-        if not isinstance(error, jsonschema.exceptions.ValidationError):
-            continue
-
-        if error.validator != "const":
-            continue
-
-        if error.path[-1] != "data_format":
-            continue
-
-        if error.instance != artifact["data_format"]:
-            continue
-
-        # Since data_format is specified as a constant in the schema,
-        # the validator_value on this exception will be the desired data format.
-        artifact["data_format"] = error.validator_value
-        return
-
-    # data format was not set!
-
-
-def _get_uuid_info(ct: dict, artifact_uuid: str) -> (dict, dict):
-    # Using uuids to find path in CT where corresponding artifact is located. Uuids are unique.
-    uuid_field_path = get_path(ct, artifact_uuid)
-
-    # As "uuid_field_path" contains path to a field with uuid,
-    # we're looking for an artifact that contains it, not the "string" field itself
-    # That's why we need skip_last=1, to get 1 "level" higher
-    # from 'uuid_field_path' field to it's parent - existing_artifact obj
-    return get_source(ct, uuid_field_path, skip_last=1)
-
-
-def merge_artifact(
-    ct: dict,
-    artifact_uuid: str,
-    object_url: str,
-    assay_type: str,
-    file_size_bytes: int,
-    uploaded_timestamp: str,
-    crc32c_hash: Optional[str] = None,
-    md5_hash: Optional[str] = None,
-) -> (dict, dict, dict):
-    """
-    create and merge an artifact into the metadata blob
-    for a clinical trial. The merging process is automatically
-    determined by inspecting the gs url path.
-    Args:
-        ct: clinical_trial object to be searched
-        artifact_uuid: artifact identifier
-        object_url: the gs url pointing to the object being added
-        file_size_bytes: integer specifying the number of bytes in the file
-        uploaded_timestamp: time stamp associated with this object
-        md5_hash: md5 hash of the uploaded object, provided by GCS for non-composite objects
-        crc32c_hash: crc32c hash of the uploaded object, provided by GCS for all objects
-    Returns:
-        ct: updated clinical trial object
-        artifact: updated artifactf
-        additional_artifact_metadata: relevant metadata collected while updating artifact
-    """
-    assert (
-        crc32c_hash or md5_hash
-    ), f"Either crc32c_hash or md5_hash must be provided for artifact: {object_url}"
-
-    # urls are created like this in _process_property:
-    file_name, uuid = object_url.split("/")[-2:]
-
-    artifact_patch = {
-        # TODO 1. this artifact_category should be filled out during prismify
-        "artifact_category": "Assay Artifact from CIMAC",
-        "object_url": object_url,
-        "file_name": file_name,
-        "file_size_bytes": file_size_bytes,
-        "uploaded_timestamp": uploaded_timestamp,
-    }
-
-    if crc32c_hash:
-        artifact_patch["crc32c_hash"] = crc32c_hash
-    if md5_hash:
-        artifact_patch["md5_hash"] = md5_hash
-
-    return _update_artifact(ct, artifact_patch, artifact_uuid)
-
-
-class InvalidMergeTargetException(ValueError):
-    """Exception raised for target of merge_clinical_trial_metadata being non schema compliant."""
-
-
-def merge_artifact_extra_metadata(
-    ct: dict, artifact_uuid: str, assay_hint: str, extra_metadata_file: BinaryIO
-) -> (dict, dict, dict):
-    """
-    Merges parsed extra metadata returned by extra_metadata_parsing to
-    corresponding artifact objects within the patch.
-    Args:
-        ct: preliminary patch from upload_assay
-        artifact_uuid: passed from upload assay
-        assay_hint: assay type
-        extra_metadata_file: extra metadata file in BinaryIO
-    Returns:
-        ct: updated clinical trial object
-        artifact: updated artifact
-        additional_artifact_metadata: relevant metadata collected while updating artifact
-    """
-
-    if assay_hint not in _EXTRA_METADATA_PARSERS:
-        raise Exception(f"Assay {assay_hint} does not support extra metadata parsing")
-    extract_metadata = _EXTRA_METADATA_PARSERS[assay_hint]
-
-    artifact_extra_md_patch = extract_metadata(extra_metadata_file)
-    return _update_artifact(ct, artifact_extra_md_patch, artifact_uuid)
-
-
-def _wes_pipeline_config(
-    upload_type: str,
-) -> Callable[[dict, dict, str], Dict[str, str]]:
-    """
-        Returns a function that for specified assay type will
-        generate a snakemake .yaml for each tumor/normal pair with
-        .fastq or bam available. 
-
-        upload_type == "assay" is for when the upload consists of .fastq/.bam assay data.
-        So sample_id's for which we now have data should be in the `patch` arg,  
-        and corresponding sample pairing info should be in the `full_ct` doc.
-
-        Or if upload_type == "pairing" it's vice versa - new runs/pairs are calculated 
-        from `patch` (which is a "pairing"), and then we look for sample data info in 
-        the `full_ct`  
-
-        In brush strokes it goes like this:
-        - we need to figure out what new data we've got
-        - based on that we calculate candidate analysis runs (tumor/normal pairs),
-          that we might need to run now, due to new data arived.
-        - last step actually check if for each candidate we have assay data
-          for both - normal and tumor sample - and we render pipeline configs for those.
-
-        As a result each assay or tumor_normal_pairing upload that "completes" some
-        analysis run/pair (making all 3 pieces available - both samples data and a pairing)
-        will be rendered only once - right after that upload.
-
-        Patch is expected to be already merged into full_ct.
-    """
-    if upload_type not in ["assay", "pairing"]:
-        raise NotImplementedError(
-            f"Not supported type:{upload_type} for wes pipeline config generation."
-        )
-
-    templ = load_pipeline_config_template("wes_analysis_config")
-
-    class AnalysisRun(NamedTuple):
-        normal_cimac_id: str
-        tumor_cimac_id: str
-        run_id: str
-
-    def internal(full_ct: dict, patch: dict, bucket: str) -> Dict[str, str]:
-        """
-            Generates a map from analysis run_ids found in full_ct
-            to generated snakemake wes .yaml configs. 
-
-            Patch is expected to be already merged into full_ct.
-        """
-
-        potential_new_runs: List[AnalysisRun] = []  # to be rendered
-
-        if upload_type == "assay":
-            # first we search for cimac_ids for which we just got new data files
-            new_data_ids = set()
-            # as we know that `patch` is a prism result of a wes upload
-            # we are sure these getitem calls should be fine
-            for wes in patch["assays"]["wes"]:
-                for r in wes["records"]:
-                    new_data_ids.add(r["cimac_id"])
-
-            # then we compose a list of all the analysis runs
-            # which are also cimac_ids tumor/normal pairs.
-            for r in (
-                full_ct.get("analysis", {})
-                .get("wes_analysis", {})
-                .get("pair_runs", [])
-                # this notation is used due to that we don't know if we have any "analyses"
-                # or any pairs in them. Thus we provide default empty containers.
-            ):
-                norm_i = r["normal"]["cimac_id"]
-                tum_i = r["tumor"]["cimac_id"]
-                runi = r["run_id"]
-                # we filter runs, for which we have new data for at least on of the samples:
-                if norm_i in new_data_ids or tum_i in new_data_ids:
-                    potential_new_runs.append(AnalysisRun(norm_i, tum_i, runi))
-
-        if upload_type == "pairing":
-            # first we filter cimac_ids for which we now got pairing info
-            # from analysis runs.
-            potential_new_runs = [
-                AnalysisRun(
-                    r["normal"]["cimac_id"], r["tumor"]["cimac_id"], r["run_id"]
-                )
-                for r in patch["analysis"]["wes_analysis"]["pair_runs"]
-            ]
-
-        # then we compose a list of all records from all assay runs,
-        # so we can filter out analysis runs for which we have data for both samples
-        all_wes_records = {}
-        for wes in full_ct["assays"]["wes"]:
-            for r in wes["records"]:
-                all_wes_records[r["cimac_id"]] = r
-
-        res = {}
-        for run in potential_new_runs:
-            # if we have data files for *both* items in a tumor/normal pair
-            if (
-                run.normal_cimac_id in all_wes_records
-                and run.tumor_cimac_id in all_wes_records
-            ):
-                # then this is a run we need, and so we render it
-                res[run.run_id + ".yaml"] = templ.render(
-                    **{
-                        "run_id": run.run_id,
-                        "tumor_sample": all_wes_records[run.normal_cimac_id],
-                        "normal_sample": all_wes_records[run.tumor_cimac_id],
-                        "BIOFX_BUCKET_NAME": bucket,
-                    }
-                )
-
-        return res
-
-    return internal
-
-
-def _rnaseq_pipeline_config(full_ct: dict, patch: dict, bucket: str) -> Dict[str, str]:
-    """
-        Generates a .yaml config for RNAseq pipeline 
-        
-        Patch is expected to be already merged into full_ct.
-    """
-
-    templ = load_pipeline_config_template("rnaseq_analysis_config")
-
-    # as we know that `patch` is a prism result of a rna_fastq upload
-    # we are sure these getitem calls should be fine
-    new_data = [r for assay in patch["assays"]["rna"] for r in assay["records"]]
-
-    res = {}
-    for sample_data in new_data:
-        res[sample_data["cimac_id"] + ".yaml"] = templ.render(
-            BIOFX_BUCKET_NAME=bucket, **sample_data
-        )
-
-    return res
-
-
-# This is a map from a assay type to a config generators,
-# that should take (full_ct: dict, patch: dict, bucket: str) as arguments
-# and return a map {"file_name": ["whatever pipeline config is"]}
-_ANALYSIS_CONF_GENERATORS = {
-    "wes_fastq": _wes_pipeline_config("assay"),
-    "wes_bam": _wes_pipeline_config("assay"),
-    "tumor_normal_pairing": _wes_pipeline_config("pairing"),
-    "rna_fastq": _rnaseq_pipeline_config,
-    "rna_bam": _rnaseq_pipeline_config,
-}
-
-
-def generate_analysis_configs_from_upload_patch(
-    ct: dict, patch: dict, template_type: str, bucket: str
-) -> Dict[str, str]:
-    """
-    Generates all needed pipeline configs, from a new upload info.
-    Args:
-        ct: full metadata object *with `patch` merged*!
-        patch: metadata patch passed from upload (assay or manifest)
-        template_type: assay or manifest type
-        bucket: a name of a bucket where data files are expected to be 
-                available for the pipeline runner
-    Returns:
-        Filename to pipeline configs as a string map.
-    """
-
-    if template_type not in _ANALYSIS_CONF_GENERATORS:
-        return {}
-
-    return _ANALYSIS_CONF_GENERATORS[template_type](ct, patch, bucket)
-
-
-def _update_artifact(
-    ct: dict, artifact_patch: dict, artifact_uuid: str
-) -> (dict, dict, dict):
-    """ Updates the artifact with uuid `artifact_uuid` in `ct`,
-    and return the updated clinical trial and artifact objects
-    Args:
-        ct: clinical trial object
-        artifact_patch: artifact object patch
-        artifact_uuid: artifact identifier
-    Returns:
-        ct: updated clinical trial object
-        artifact: updated artifact
-        additional_artifact_metadata: relevant metadata collected while updating artifact
-    """
-
-    artifact, additional_artifact_metadata = _get_uuid_info(ct, artifact_uuid)
-
-    # TODO this might be better with merger:
-    # artifact_schema = load_and_validate_schema(f"artifacts/{artifact_type}.json")
-    # artifact_parent[file_name] = Merger(artifact_schema).merge(existing_artifact, artifact)
-    artifact.update(artifact_patch)
-
-    _set_data_format(ct, artifact)
-
-    # return the artifact that was merged and the new object
-    return ct, artifact, additional_artifact_metadata
-
-
-class MergeCollisionException(ValueError):
-    pass
-
-
-class ThrowOnCollision(strategies.Strategy):
-    """
-    Similar to the jsonmerge's built in 'discard' strategy,
-    but throws an error if the value already exists. This is
-    used to prevent updates in `merge_clinical_trial_metadata`.
-    """
-
-    def merge(self, walk, base, head, schema, meta, **kwargs):
-        if base.is_undef():
-            return head
-        if base.val != head.val:
-            prop = base.ref.rsplit("/", 1)[-1]
-            raise MergeCollisionException(
-                f"Found conflicting values for {prop}: {base.val} (current) != {head.val} (incoming). "
-                "Updates are not currently supported."
-            )
-        return base
-
-    def get_schema(self, walk, schema, **kwargs):
-        return schema
-
-
-PRISM_MERGE_STRATEGIES = {
-    # This overwrites the default jsonmerge merge strategy for literal values.
-    "overwrite": ThrowOnCollision(),
-    # Alias the builtin jsonmerge overwrite strategy
-    "overwriteAny": strategies.Overwrite(),
-}
-
-
-def merge_clinical_trial_metadata(patch: dict, target: dict) -> (dict, List[str]):
-    """
-    Merges two clinical trial metadata objects together
-    Args:
-        patch: the metadata object to add
-        target: the existing metadata object
-    Returns:
-        arg1: the merged metadata object
-        arg2: list of validation errors
-    """
-
-    validator = load_and_validate_schema("clinical_trial.json", return_validator=True)
-
-    # first we assert original object is valid
-    try:
-        validator.validate(target)
-    except jsonschema.ValidationError as e:
-        raise InvalidMergeTargetException(
-            f"Merge target is invalid: {target}\n{e}"
-        ) from e
-
-    # assert the un-mutable fields are equal
-    # these fields are required in the schema
-    # so previous validation assert they exist
-    if patch.get(PROTOCOL_ID_FIELD_NAME) != target.get(PROTOCOL_ID_FIELD_NAME):
-        raise InvalidMergeTargetException(
-            "Unable to merge trials with different " + PROTOCOL_ID_FIELD_NAME
-        )
-
-    # merge the two documents
-    merger = Merger(validator.schema, strategies=PRISM_MERGE_STRATEGIES)
-    merged = merger.merge(target, patch)
-
-    return merged, list(validator.iter_errors(merged))
-
-
-# Build a regex from the CIMAC ID pattern in the schema
-cimac_id_regex = re.compile(
-    load_and_validate_schema("sample.json")["properties"]["cimac_id"]["pattern"]
-)
-
-
-def parse_elisa(xlsx: BinaryIO) -> dict:
-    """
-    Parses the given ELISA grand serology results file to extract a list of sample IDs.
-    If the file is not valid NPX but still xlsx the function will
-    return a dict containing an empty list. Sample IDs not conforming to the CIMAC ID
-    format will be skipped. The function will pass along any IO errors.
-    Args:
-        xlsx: an opened NPX file
-    Returns:
-        arg1: a dict of containing list of sample IDs and number of samples
-    """
-
-    # load the file
-    if type(xlsx) == str:
-        raise TypeError(f"parse_npx only accepts BinaryIO and not file paths")
-
-    workbook = openpyxl.load_workbook(xlsx)
-
-    # extract data to python
-    ids = []
-    worksheet = workbook[workbook.sheetnames[0]]
-
-    for i, row in enumerate(worksheet.iter_rows()):
-
-        if i == 0:
-            assert "CIMAC ID" == row[0].value
-            continue
-
-        val = row[0].value
-
-        if val:
-            if cimac_id_regex.match(val):
-                ids.append(val)
-
-    sample_count = len(ids)
-
-    samples = {"samples": ids, "number_of_samples": sample_count}
-
-    return samples
-
-
-def parse_npx(xlsx: BinaryIO) -> dict:
-    """
-    Parses the given NPX file from olink to extract a list of sample IDs.
-    If the file is not valid NPX but still xlsx the function will
-    return a dict containing an empty list. Sample IDs not conforming to the CIMAC ID
-    format will be skipped. The function will pass along any IO errors.
-    Args:
-        xlsx: an opened NPX file
-    Returns:
-        arg1: a dict of containing list of sample IDs and number of samples
-    """
-
-    # load the file
-    if type(xlsx) == str:
-        raise TypeError(f"parse_npx only accepts BinaryIO and not file paths")
-
-    workbook = openpyxl.load_workbook(xlsx)
-
-    # extract data to python
-    ids = []
-    for worksheet_name in workbook.sheetnames:
-
-        # simplify.
-        worksheet = workbook[worksheet_name]
-        seen_onlinkid = False
-        for i, row in enumerate(worksheet.iter_rows()):
-
-            # extract values from row
-            vals = [col.value for col in row]
-
-            first_cell = vals[0]
-
-            # skip empty
-            if len(vals) == 0 or first_cell is None:
-                continue
-
-            # check if we are starting ids
-            if first_cell == "OlinkID":
-                seen_onlinkid = True
-                continue
-
-            # check if we are done.
-            if first_cell == "LOD":
-                break
-
-            # get the identifier
-            if seen_onlinkid:
-                # check that it is a CIMAC ID
-                if cimac_id_regex.match(first_cell):
-                    ids.append(first_cell)
-
-    sample_count = len(ids)
-
-    samples = {"samples": ids, "number_of_samples": sample_count}
-
-    return samples
-
-
-_EXTRA_METADATA_PARSERS = {"olink": parse_npx, "elisa": parse_elisa}
